@@ -9,11 +9,12 @@
   },
   "dependencies": {
     "@mswertz/emx2-styleguide": "*",
-    "core-js": "3.15.2",
+    "@vue/compat": "3.1.0",
+    "core-js": "3.6.5",
     "graphql-request": "3.4.0",
-    "vue": "2.6.14",
-    "vue-router": "3.5.2",
-    "vuex": "3.6.2"
+    "vue": "3.1.0",
+    "vue-router": "4.0.3",
+    "vuex": "4.0.0"
   },
   "eslintConfig": {
     "root": true,
