@@ -1,14 +1,11 @@
-import Vue from "vue";
-import Vuex from "vuex";
+import { createStore as _createStore } from "vuex";
 
 import actions from "./actions";
 import getters from "./getters";
 import mutations from "./mutations";
 import state from "./state";
 
-Vue.use(Vuex);
-
-export default new Vuex.Store({
+export default _createStore({
   actions,
   getters,
   mutations,
